sets up to handle the `ec` from the event
special attention is put to make sure that it is done ahead of the event
firing off.  Also it will give you the option to set an action just
after the data is set as needed.  There is no protect to a
missconfiguration as it is expected that you use the Google debuger
since we are just basically proxy'ing the settings per
https://developers.google.com/analytics/devguides/collection/analyticsjs/enhanced-ecommerce