`ev` (event value) must not be a empty string or null
if you pass in a null or an empty string you get this warning.
`Expected a number value for the field: "eventValue". but found:
"string".` We made the change to null to address #52 but setting it to
null was the wrong choice as it kills the event per the warning from
google.  This is the correct way, which is default to 0 per docs